requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
PyMySQL==1.1.1
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
httpx==0.28.1
numpy==2.2.5


certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
six==1.17.0
typing_extensions==4.13.0
tzdata==2025.2
